scsi: reduce retries to 1
When errors are reported to the SCSI layer, there tend to have been retries
performed by the SAN or HDD reporting the error.  Since we make use of
multipath on SANs, and expect drive retries via ATA, reduce the number of
retries caused by the SCSI layer to 1 to allow other parts of the system to
try to cope with the error in a timely fashion.