fs/ioctl.c: add support for FIGETRESBYTES64 ioctl
Add support for the FIGETRESBYTES64 ioctl which gets the number of bytes
of file data resident in the page cache for a given file.  This can be
used as a quick and low overhead way of determining if most of a file is
cached when making decisions about scheduling reads of a file.

(cherry picked from commit aa3ada8ed71d6be25c00411c4ffb67d88370e0e6)