ipmi/ipmi_poweroff: Use a work queue to avoid lock contention
When ipmi_poweroff is built-in it will use a smi_watcher
to initialize itself, but during initialization the ipmi_ssif interface
hasn't release all the locks it holds and prevents ipmi_create_user
from creating a new user because of the locks.  Therefore a work
queue was added to delay this work until initialization is complete.