Solace: ensure that file descriptors are closed early during core dumps
When a multi-gigabyte process dumps core it can take a substantial
amount of time to write the core file to disk.  During this time, all
file descriptors belonging to the process are held open.  This prevents
any filesystems the process was accessing from being umount()ed.  This
behaviour negatively effects failover time for the appliance.

Fix this by making fput() use the delayed work code path so that file
descriptor release is performed independant of the core dump process.

Signed-off-by: Benjamin LaHaise <bcrl@kvack.org>

Conflicts:
	fs/file_table.c