Solace: ata: Add failfast support, and command timeout fault injection
Adding support for the upper layer to request the disk to failfast during error
recovery instead of doing resets.  This is important for instances where delays
in trying recover cause a worse issue then disabling the disk.

When fault injection is enable through debugfs commands to the disk will not
be sent to hardware and simulate a communication failure with the disk.  The
number entered for the failure is how many times check ready will fail,
preventing the disk been ready after a reset.  This injected fault very closely
simulates how disks fail.

Conflicts:

	drivers/ata/ahci.c
	drivers/ata/libahci.c
	include/linux/libata.h

Conflicts:
	drivers/ata/libata-eh.c
	drivers/ata/libata-scsi.c
	include/linux/libata.h