Solace/watchdog: Add a flag to panic now in the hard lockup detector
Solace requires a method for pci devices to quickly restart the
host. A bonus is having a power cycle and the reason for
the restart logged.  This restart will happen when the host may be
hung.  PCI device can write to physical memory so updating a flag
in memory is a simple solution.

Conflicts:

	kernel/watchdog.c

Conflicts:
	kernel/watchdog_perf.c
	lib/Kconfig.debug