qla2xxx: add fake_pause support and disable firmware dump alloc by default
Add fake_pause support to qla2xxx for simulating pauses.

Disabing firmware dump removes timeouts that are seen while firmware dumping.

Also this is a Solace local change to cause the qla2xxx driver to panic() the
kernel on the 6th firmware dump attempt within 15 minutes.  This was added
to mitigate the failure seen in bug 42768.

Conflicts:

	drivers/scsi/qla2xxx/qla_def.h

 Conflicts:
	drivers/scsi/qla2xxx/qla_def.h