x86/nmi: Show state filter with support to not stack dump
On a NMI cause by a IPMI watchdog pre timer it is use full to get
a show state, however IPMI SEL records are limited therefore it is required to
keep logs small.  An option was added to dump the stack only
on tasks that are in as specific state (i.e. TASK_UNINTERRUPTIBLE).

Conflicts:

	kernel/sched/core.c

Conflicts:
	include/linux/sched.h
	kernel/sched/core.c