char/ipmi: Improvements of ipmi sel kmsg dump during NMIs
Some motherboards support an IPMI watchdog pre NMI, that will trigger
x number of seconds before the watchdog causes a reboot.  This limited
time means kmsg dump needs to write to the sel as soon as possible in an
nmi.  Also if in a NMI, then powercycle so the kdump kernel isn't run.

Conflicts:

	drivers/char/ipmi/ipmi_msghandler.c