Solace: bug 44953: workaround for b_dirty = 1 after umount() and sync
With the added checks in bug 44953 to examine if there are any dirty buffers
outstanding on a block device after the final umount() and sync, the b_dirty
stat was showing up with a count of 1.  The inode outstanding as reported by
this count is always the underlying block device's inode, which never seems
to have any actual dirty buffers outstanding on it.

In order to keep the check for dirty inodes outstanding on the block device,
modify the b_dirty stat reporting to suppress the block device inode when
it is outstanding on the b_dirty list with no dirty buffers.

Conflicts:

	mm/backing-dev.c

(cherry picked from commit e072ae2f39a89d17a69841e3ba44fc6229e08df6)